Removed TransformServiceRegistryConfigTest (contained 2 failing tests)
It contained two tests that now fail as we now correctly don't allow single step transformers in pipeline files that don't overlay a t-engine transforms. We already test this in the alfresco-transform-model. Once removed all the remaining code was moved to the sub class LocalTransformServiceRegistryConfigTest.
